fet: dropdown 관 component align 수정
