+++ src/data/components/dropdown.js
@@ -16,3 +16,23 @@ export const FIELD = [
     { id: 5, text: "광주" },
   ];
   
+
+  export const DUMMY_DEPART = [
+    { id: 0, text: "사회과학대학" },
+    { id: 1, text: "미술대학" },
+    { id: 2, text: "융합인문대학" },
+    { id: 3, text: "미래공과대학" },
+    { id: 4, text: "경영경제대학" },
+    { id: 5, text: "공연예술대학" },
+  ];
+  
+  export const DUMMY_MAJOR = [
+    { id: 0, text: "사회학과" },
+    { id: 1, text: "행정학과" },
+    { id: 2, text: "정치외교학과" },
+    { id: 3, text: "금속공예학과" },
+    { id: 4, text: "목조형가구디자인과" },
+    { id: 5, text: "산업디자인과" },
+    { id: 6, text: "예술학과" },
+    { id: 7, text: "예술경영학과" },
+  ];
