+++ src/data/components/sortTag.js
@@ -2,7 +2,7 @@ export const SORT_TAG_DATA = [
   { id: 0, content: "전체", select: true },
   { id: 1, content: "단과대 동아리", select: false },
   { id: 2, content: "과 동아리", select: false },
-  { id: 2, content: "기타 교내 동아리", select: false },
-  { id: 2, content: "중앙 동아리", select: false },
-  { id: 3, content: "연합 동아리", select: false },
+  { id: 3, content: "기타 교내 동아리", select: false },
+  { id: 4, content: "중앙 동아리", select: false },
+  { id: 5, content: "연합 동아리", select: false },
 ];
